pyproject.toml
[project]
name = "explainable-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
	"pandas>=2.2.0",
	"numpy>=1.24,<2",
	"numba>=0.59,<0.60",
	"llvmlite>=0.42,<0.43",
	"scikit-learn>=1.4.0",
	"shap>=0.44.0",
	"streamlit>=1.38.0",
	"xgboost>=2.0.0",
	"joblib>=1.3.0",
	"PyYAML>=6.0",
]
